CheckCollisionCircleRec isn't working
other than that collision is half working, need to rewrite collision for triangle since currently it just checks a small circle instead of actually finding the 3 points of the triangle. Will likely need raymath to get the points of the triangle. Game is however, very un fun and I am not sure this is even worth finishing